Restructuring - formalizing definitions
