+++ Source/Kernel/Events.Projections/Definitions/ChildrenDefinition.cs
@@ -1,22 +1,20 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Cratis.Events.Projections.Json
+namespace Cratis.Events.Projections.Definitions
 {
     /// <summary>
-    /// Represents the definition of a projection.
+    /// Represents the definition of a children projection.
     /// </summary>
-    /// <param name="Identifier"><see cref="ProjectionId">Identifier</see> of the projection.</param>
-    /// <param name="Name">Friendly displayname of the projection.</param>
+    /// <param name="IdentifiedBy">Property on model that identifies the unique object, typically the key - or id (event source id).</param>
     /// <param name="Model">The target <see cref="ModelDefinition"/>.</param>
     /// <param name="From">All the <see cref="FromDefinition"/> for <see cref="EventType">event types</see>.</param>
     /// <param name="Children">All the <see cref="ChildrenDefinition"/> for properties on model.</param>
     /// <param name="RemovedWith">The definition of what removes a child, if any.</param>
-    public record ProjectionDefinition(
-        ProjectionId Identifier,
-        ProjectionName Name,
+    public record ChildrenDefinition(
+        Property IdentifiedBy,
         ModelDefinition Model,
         IDictionary<EventType, FromDefinition> From,
         IDictionary<Property, ChildrenDefinition> Children,
-        RemovedWithDefinition? RemovedWith);
+        RemovedWithDefinition? RemovedWith) : ProjectionDefinition(Guid.Empty, string.Empty, Model, From, Children, RemovedWith);
 }
