+++ Source/Kernel/Events.Projections/Definitions/ModelDefinition.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Cratis.Events.Projections.Json
+namespace Cratis.Events.Projections.Definitions
 {
     /// <summary>
     /// Represents the definition of a model used by a projection.
