+++ Source/Kernel/Events.Projections/Json/PropertyChildrenDefinitionDictionaryJsonConverter.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using Newtonsoft.Json;
+using Cratis.Events.Projections.Definitions;
 
 namespace Cratis.Events.Projections.Json
 {
